requirements.txt
numpy
matplotlib
opencv-python
deepface
flask
streamlit_modal